requirements.txt
ipykernel==6.30.1
numpy==1.26.4
pandas==2.3.2
scikit-learn==1.7.1
matplotlib==3.10.5
plotly==6.3.0
seaborn==0.13.2
scipy==1.16.1
imbalanced-learn==0.14.0
imblearn==0.0
botocore==1.40.16
fastapi==0.116.1
uvicorn==0.35.0
jinja2==3.1.6
python-multipart==0.0.20
PyYAML==6.0.2
tqdm==4.67.1
joblib==1.5.1
types-PyYAML==6.0.12.20250822
streamlit==1.48.1
notebook==7.4.5
jupyterlab==4.4.6
catboost==1.2.8
xgboost==3.0.4
python-dotenv==1.1.1
pymongo==4.14.1
from-root==1.3.0
evidently==0.2.8
mlflow==3.3.1
dill==0.4.0
pandera
-e .